Agglomération des fichiers de test + nouveau test
Tout les scripts de test sont dans le dossier script/test.

test_table ne test pas la table mais les fonctionnalités qu'eel doit
implémenter ne plus l'utiliser.
A la place utiliser un objet CarcasheimBack bouchonnat les
fonctionnalités du displaysystem.

+++ Carcassheim_unity/Assets/Affichage_InGame/TestScripts/test_player_list.cs
@@ -31,7 +31,7 @@ public class test_player_list : MonoBehaviour
                     banner.setPlayer(pl);
                 }
                 // Debug.Log("Player nb : " + ((uint)pl_list.nbPlayer()).ToString());
-                banner.setPlayerNumber((uint)pl_list.nbPlayer());
+                banner.setPlayerNumber(pl_list.nbPlayer());
             }
             else if (Input.GetKeyUp("n"))
             {
